Refactor the code to change the "register" dialog for "enter_license" dialog
