+++ data/strings/en.ini
@@ -422,7 +422,7 @@ PlayAnimation = Play Animation
 PlayPreviewAnimation = Play Preview Animation
 Redo = Redo
 Refresh = Refresh
-Register = Register Aseprite
+EnterLicense = Enter License
 RemoveFrame = Remove Frame
 RemoveFrameTag = Remove Tag
 RemoveLayer = Remove Layer
@@ -1044,7 +1044,7 @@ help_documentation = Documentation
 help_tutorial = Tutorial
 help_release_notes = Release Notes
 help_twitter = Twitter
-help_register = &Register Aseprite
+help_enter_license = Enter &License
 help_about = &About
 
 [modify_selection]
@@ -1063,10 +1063,10 @@ name = Name:
 tileset = Tileset:
 default_new_layer_name = New Layer
 
-[register]
-title = Register Aseprite
-default_message = If you need an activation code, go to
-activation_code = Activation Code
+[enter_license]
+title = Enter License
+default_message = If you need a license key, go to
+license_key = License Key
 
 [tileset_selector]
 new_tileset = New Tileset
