+++ data/widgets/enter_license.xml
@@ -1,22 +1,18 @@
 <!-- Aseprite -->
 <!-- Copyright (c) 2021  Igara Studio S.A. -->
 <gui>
-  <window id="register" text="@.title" expansive="false">
+  <window id="enter_license" text="@.title" expansive="false">
     <box vertical="true">
       <hbox>
         <label text="@.default_message" />
         <link text="https://www.aseprite.org/download/" url="https://www.aseprite.org/download/" />
       </hbox>
-      <separator text="@.activation_code" left="true" horizontal="true" />
-      <grid columns="4">
-        <entry id="code1" cell_align="horizontal" magnet="true" maxsize="5" />
-        <entry id="code2" cell_align="horizontal" maxsize="5" />
-        <entry id="code3" cell_align="horizontal" maxsize="5" />
-      </grid>
+      <separator text="@.license_key" left="true" horizontal="true" />
+      <entry id="license_key" cell_align="horizontal" magnet="true" maxsize="17" />
       <box horizontal="true">
         <box horizontal="true" expansive="true" />
         <box horizontal="true" homogeneous="true">
-          <button text="@general.ok" closewindow="true" id="ok_button" magnet="true" width="60" />
+          <button text="@general.ok" id="ok_button" magnet="true" width="60" />
           <button text="@general.cancel" closewindow="true" />
         </box>
       </box>
